requirements.txt
flask>=2.3.0
gradio>=4.0.0
fastapi>=0.115.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.6.0
numpy>=1.24.0
python-docx>=1.1.0
